requirements.txt.txt
python-telegram-bot[job-queue]
psycopg2-binary
flask
gunicorn
groq